src/embed/weaviate/requirements.txt
weaviate-client==4.9.0
pymongo==4.6.1
python-dotenv==1.0.1
google-generativeai==0.7.2
google-ai-generativelanguage==0.6.6
httpx==0.27.0
numpy==1.26.4
tenacity==8.2.3